requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
PyPDF2==3.0.1
pdfplumber==0.9.0
openai==1.3.5
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.8.0
email-validator==2.0.0
gunicorn==21.2.0